split tests for geolocation vs sensor
